Add ContainerEntityListenerResolver + load listeners lazy by default
