+++ Tests/DependencyInjection/Fixtures/config/yml/orm_entity_listener_resolver.yml
@@ -1,14 +1,14 @@
 services:
     entity_listener_resolver:
-        class: \Doctrine\ORM\Mapping\DefaultEntityListenerResolver
+        class: Doctrine\ORM\Mapping\DefaultEntityListenerResolver
 
     entity_listener1:
-        class: \EntityListener
+        class: EntityListener
         tags:
             - { name: doctrine.orm.entity_listener }
 
     entity_listener2:
-        class: \EntityListener
+        class: EntityListener
         tags:
             - { name: doctrine.orm.entity_listener, entity_manager: em2 }
